refactor: move queryInformationSchema into spanner.go
